+++ packages/auth-client/src/db.ts
@@ -1,5 +1,5 @@
 import { openDB, IDBPDatabase } from 'idb';
-import { isBrowser, KEY_STORAGE_DELEGATION, KEY_STORAGE_KEY } from './storage';
+import { DB_VERSION, isBrowser, KEY_STORAGE_DELEGATION, KEY_STORAGE_KEY } from './storage';
 
 type Database = IDBPDatabase<unknown>;
 type IDBValidKey = string | number | Date | BufferSource | IDBValidKey[];
@@ -69,7 +69,7 @@ export class IdbKeyVal {
    * @constructs an {@link IdbKeyVal}
    */
   public static async create(options?: DBCreateOptions): Promise<IdbKeyVal> {
-    const { dbName = AUTH_DB_NAME, storeName = OBJECT_STORE_NAME, version = 1 } = options ?? {};
+    const { dbName = AUTH_DB_NAME, storeName = OBJECT_STORE_NAME, version = DB_VERSION } = options ?? {};
     const db = await _openDbStore(dbName, storeName, version);
     return new IdbKeyVal(db, storeName);
   }